Major Fix: Check if the user is in a guild and then remove all of those changes.
I only execute guild functions when a person is in a guild.

----------------------

Last update showed that it doesn't do guild AND whisper anymore but it does. This is because the whisper is every cycle but guild is every 2 or 3. (3 technically, but timer depends on Pantheon servers).

I need to do something else like store them in a thing and set a timer and if that name doesn't get added for a set timer or # of cycles, then it logs them as online or off, etc. Not a big deal, just a small double tap.